Add exceptions to class

+++ StereoMiddleburyEval/StereoMiddleburyEval/__init__.py
@@ -7,9 +7,54 @@ import wget
 import sys
 import zipfile
 
+# define StereoMiddleburyEval exceptions
+class InvalidSceneName(Exception):
+    def __str__(self):
+        return "Invalid scene name. Use getSceneList() to get full list of scenes available"
+
+class MalformedPFM(Exception):
+    def __init__(self, message="Malformed PFM file"):
+        self.message = message
+        super().__init__(self.message)
+
+    def __str__(self):
+        return self.message
+
+class PathNotFound(Exception):
+    def __init__(self, filename, message="Path does not exist: "):
+        self.filename = filename
+        self.message = message
+        super().__init__(self.message)
+
+    def __str__(self):
+        return self.message+self.filename
+
 class StereoMiddleburyEval():
     def __init__(self):
-        pass
+        self.scene_list = [
+            "Adirondack",
+            "Backpack",
+            "Bicycle1",
+            "Cable",
+            "Classroom1",
+            "Couch",
+            "Flowers",
+            "Jadeplant",
+            "Mask",
+            "Motorcycle",
+            "Piano",
+            "Pipes",
+            "Playroom",
+            "Playtable",
+            "Recycle",
+            "Shelves",
+            "Shopvac",
+            "Sticks",
+            "Storage",
+            "Sword1",
+            "Sword2",
+            "Umbrella",
+            "Vintage"]
 
     def disp_to_depth(self,disp,cal_filepath):
         #Sample calib.txt file for one of the full-size training image pairs:
@@ -30,7 +75,7 @@ class StereoMiddleburyEval():
         if (not os.path.exists(cal_filepath)):
             print("Calibration file not found")
             print(cal_filepath)
-            raise Exception("Calibration file not found")
+            raise PathNotFound(cal_filepath,"Calibration file not found")
         
         # Open calibration file
         file = open(cal_filepath, 'rb')
@@ -94,14 +139,14 @@ class StereoMiddleburyEval():
             elif header == 'Pf':
                 color = False
             else:
-                raise Exception('Not a PFM file.')
+                raise MalformedPFM('Not a PFM file.')
             
             # Read dimensions from pfm file and check they match expected
             dim_match = re.match(r'^(\d+)\s(\d+)\s$', file.readline().decode('utf-8'))
             if dim_match:
                 width, height = map(int, dim_match.groups())
             else:
-                raise Exception('Malformed PFM header.')
+                raise MalformedPFM('Malformed PFM header')
                 
             # Read data scale from file
             scale = float(file.readline().decode('utf-8').rstrip())
@@ -120,86 +165,68 @@ class StereoMiddleburyEval():
             data = cv2.flip(data,0)
             return data, scale
         else:
-            raise Exception('File does not exist')
+            raise PathNotFound(filepath,"Pfm file does not exist")
 
-    def load_scene_data(self,scene_folder,display_images=False):
+    def load_scene_data(self,scene_name,dataset_folder,display_images=False):
         # Load scene data from scene folder
+        if (self.checkValidSceneName(scene_name)):
+            display_time = 500
+            # Get name of disparity image (pfm) in folder
+            disp_filename = os.path.join(dataset_folder,scene_name,"disp0.pfm")
+            # Check disparity file exists
+            if (not os.path.exists(disp_filename)):
+                print("Disparity pfm file does not exist")
+                print(disp_filename)
+                raise PathNotFound(disp_filename,"Disparity pfm file does not exist")
+            # Load disparity file to numpy image
+            disp_image, _ = self.load_pfm(disp_filename)
+            if (display_images):
+                # Display disparity image in opencv window
+                norm_disp_image = self.normalise_pfm_data(disp_image)
+                norm_disp_image_resize = cv2.resize(norm_disp_image, dsize=(0, 0), fx=0.2, fy=0.2)
+                cv2.imshow('image', cv2.applyColorMap(norm_disp_image_resize, cv2.COLORMAP_JET))
+                cv2.waitKey(display_time)
+            
+            # Define calibration file in scene folder
+            cal_file = os.path.join(dataset_folder,scene_name,"calib.txt")
+            # Calculate depth image from disparity using calibration file
+            depth_image = self.disp_to_depth(disp_image,cal_file)
 
-        # Get name of disparity image (pfm) in folder
-        disp_filename = os.path.join(scene_folder,"disp0.pfm")
-        # Check disparity file exists
-        if (not os.path.exists(disp_filename)):
-            print("Disparity pfm file does not exist")
-            print(disp_filename)
-            raise Exception("Disparity pfm file does not exist")
-        # Load disparity file to numpy image
-        disp_image, _ = self.load_pfm(disp_filename)
-        if (display_images):
-            # Display disparity image in opencv window
-            norm_disp_image = self.normalise_pfm_data(disp_image)
-            norm_disp_image_resize = cv2.resize(norm_disp_image, dsize=(0, 0), fx=0.2, fy=0.2)
-            cv2.imshow('image', cv2.applyColorMap(norm_disp_image_resize, cv2.COLORMAP_JET))
-            cv2.waitKey(1000)
-        
-        # Define calibration file in scene folder
-        cal_file = os.path.join(scene_folder,"calib.txt")
-        # Calculate depth image from disparity using calibration file
-        depth_image = self.disp_to_depth(disp_image,cal_file)
+            # Define left and right image files in scene folder
+            left_image_filename = os.path.join(dataset_folder,scene_name,"im0.png")
+            right_image_filename = os.path.join(dataset_folder,scene_name,"im1.png")
+            # Check left and right image files exist
+            if (not os.path.exists(left_image_filename) or not os.path.exists(right_image_filename)):
+                print("Left or right image file does not exist")
+                print(left_image_filename)
+                print(right_image_filename)
+                raise PathNotFound(left_image_filename+","+right_image_filename,"Left or right image file does not exist")
+            
+            # Read left and right image files to numpy image
+            left_image = cv2.imread(left_image_filename,cv2.IMREAD_UNCHANGED)
+            right_image = cv2.imread(right_image_filename,cv2.IMREAD_UNCHANGED)
+            if (display_images):
+                # Display left and right image files to OpenCV window
+                left_image_resize = cv2.resize(left_image, dsize=(0, 0), fx=0.2, fy=0.2)
+                right_image_resize = cv2.resize(right_image, dsize=(0, 0), fx=0.2, fy=0.2)
+                cv2.imshow('image', left_image_resize)
+                cv2.waitKey(display_time)
+                cv2.imshow('image', right_image_resize)
+                cv2.waitKey(display_time)
 
-        # Define left and right image files in scene folder
-        left_image_filename = os.path.join(scene_folder,"im0.png")
-        right_image_filename = os.path.join(scene_folder,"im1.png")
-        # Check left and right image files exist
-        if (not os.path.exists(left_image_filename) or not os.path.exists(right_image_filename)):
-            print("Left or right image file does not exist")
-            print(left_image_filename)
-            print(right_image_filename)
-            raise Exception("Left or right image file does not exist")
-        
-        # Read left and right image files to numpy image
-        left_image = cv2.imread(left_image_filename,cv2.IMREAD_UNCHANGED)
-        right_image = cv2.imread(right_image_filename,cv2.IMREAD_UNCHANGED)
-        if (display_images):
-            # Display left and right image files to OpenCV window
-            left_image_resize = cv2.resize(left_image, dsize=(0, 0), fx=0.2, fy=0.2)
-            right_image_resize = cv2.resize(right_image, dsize=(0, 0), fx=0.2, fy=0.2)
-            cv2.imshow('image', left_image_resize)
-            cv2.waitKey(1000)
-            cv2.imshow('image', right_image_resize)
-            cv2.waitKey(1000)
-
-        return left_image,right_image,disp_image,depth_image
+            return left_image,right_image,disp_image,depth_image
+        raise InvalidSceneName()
 
     def getSceneList(self):
         # Return full list of scenes available
-        return [
-            "Adirondack",
-            "Backpack",
-            "Bicycle1",
-            "Cable",
-            "Classroom1",
-            "Couch",
-            "Flowers",
-            "Jadeplant",
-            "Mask",
-            "Motorcycle",
-            "Piano",
-            "Pipes",
-            "Playroom",
-            "Playtable",
-            "Recycle",
-            "Shelves",
-            "Shopvac",
-            "Sticks",
-            "Storage",
-            "Sword1",
-            "Sword2",
-            "Umbrella",
-            "Vintage"]
+        return self.scene_list
 
     def getURLfromScene(self,scene_name):
-        # Get URL on middlebury servers for 2014 dataset for chosen scene
-        return "http://vision.middlebury.edu/stereo/data/scenes2014/zip/"+scene_name+"-perfect.zip"
+        if (self.checkValidSceneName(scene_name)):
+            # Get URL on middlebury servers for 2014 dataset for chosen scene
+            return "http://vision.middlebury.edu/stereo/data/scenes2014/zip/"+scene_name+"-perfect.zip"
+        else:
+            raise InvalidSceneName()
 
     #create this bar_progress method which is invoked automatically from wget
     def bar_progress(self,current, total, width=80):
@@ -208,42 +235,47 @@ class StereoMiddleburyEval():
         sys.stdout.write("\r" + progress_message)
         sys.stdout.flush()
 
+    def checkValidSceneName(self,scene_name):
+        return scene_name in self.scene_list
+
     def downloadDataset(self,scene_name,output_folder):
-        # Check output folder exists
-        if (os.path.exists(output_folder)):
-            # Download dataset from middlebury servers
-            # Get url from scene name
-            url = self.getURLfromScene(scene_name)
-            # Get name of scene data folder
-            scene_output_folder = os.path.join(output_folder,scene_name)
-            # Define destination name for zip file
-            zip_filepath = os.path.join(output_folder,scene_name+".zip")
-            # clean-up tmp files from incomplete downloads
-            tmp_files = glob.glob(os.path.join(output_folder,"*tmp"))
-            for tmp_file in tmp_files:
-                os.remove(tmp_file)
-            # Check scene folder doesn't already exist
-            if (not os.path.exists(scene_output_folder)):
-                # Check zip file doesn't already exist
-                if (not os.path.exists(zip_filepath)):
-                    print("Downloading from: "+url)
-                    # download file from middlebury server
-                    wget.download(url, zip_filepath, bar=self.bar_progress)
+        if (self.checkValidSceneName(scene_name)):
+            # Check output folder exists
+            if (os.path.exists(output_folder)):
+                # Download dataset from middlebury servers
+                # Get url from scene name
+                url = self.getURLfromScene(scene_name)
+                # Get name of scene data folder
+                scene_output_folder = os.path.join(output_folder,scene_name)
+                # Define destination name for zip file
+                zip_filepath = os.path.join(output_folder,scene_name+".zip")
+                # clean-up tmp files from incomplete downloads
+                tmp_files = glob.glob(os.path.join(output_folder,"*tmp"))
+                for tmp_file in tmp_files:
+                    os.remove(tmp_file)
+                # Check scene folder doesn't already exist
+                if (not os.path.exists(scene_output_folder)):
+                    # Check zip file doesn't already exist
+                    if (not os.path.exists(zip_filepath)):
+                        print("Downloading from: "+url)
+                        # download file from middlebury server
+                        wget.download(url, zip_filepath, bar=self.bar_progress)
+                    else:
+                        print("Zip file for dataset already exists here, skipping download of file: "+zip_filepath)
+                    print("Extracting zip...")
+                    # unzip downloaded file
+                    with zipfile.ZipFile(zip_filepath,"r") as zip_ref:
+                        zip_ref.extractall(output_folder)
+                    # rename scene folder to remove '-perfect'
+                    os.rename(os.path.join(output_folder,scene_name+"-perfect"),scene_output_folder)
+                    # removing zip file
+                    os.remove(zip_filepath)
                 else:
-                    print("Zip file for dataset already exists here, skipping download of file: "+zip_filepath)
-                print("Extracting zip...")
-                # unzip downloaded file
-                with zipfile.ZipFile(zip_filepath,"r") as zip_ref:
-                    zip_ref.extractall(output_folder)
-                print("Organising scene folder")
-                # rename scene folder to remove '-perfect'
-                os.rename(os.path.join(output_folder,scene_name+"-perfect"),scene_output_folder)
-                # removing zip file
-                os.remove(zip_filepath)
+                    print("Dataset already exists here, skipping re-download of "+scene_name)
             else:
-                print("Dataset already exists here, skipping re-download of "+scene_name)
+                raise Exception('Output folder not found for storing datasets')
         else:
-            raise Exception('Output folder not found for storing datasets')
+            raise InvalidSceneName()
 
 if __name__ == "__main__":
     dataset_folder = os.path.join(os.getcwd(),"datasets") #Path to dowmload datasets
@@ -263,5 +295,4 @@ if __name__ == "__main__":
         stme.downloadDataset(scenename,dataset_folder) # will only download it if it hasn't already been downloaded
         # Load scene data from downloaded folder
         print("Loading data for scene '"+scenename+"'...")
-        scene_folder = os.path.join(dataset_folder,scenename)
-        left_image,right_image,disp_image,depth_image = stme.load_scene_data(scene_folder,display_images)
+        left_image,right_image,disp_image,depth_image = stme.load_scene_data(scenename,dataset_folder,display_images)
